refactor: ResourceReader가 ResourceFile을 반환하게 변경
