+++ tomcat/src/main/java/org/apache/catalina/util/ResourceFile.java
@@ -1,0 +1,4 @@
+package org.apache.catalina.util;
+
+public record ResourceFile(String contentType, String body) {
+}
