+++ tomcat/src/main/java/org/apache/coyote/http11/HttpCookie.java
@@ -4,6 +4,8 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Collectors;
+import org.apache.coyote.http11.header.HttpHeader;
+import org.apache.coyote.http11.header.HttpHeaderName;
 
 public class HttpCookie {
