+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequest.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import org.apache.coyote.http11.header.HttpHeader;
 
 public class HttpRequest {
